Moves files around, adds scss and some colors from design
