+++ src/App.js
@@ -2,9 +2,9 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { BrowserRouter as Router, Route } from 'react-router-dom';
 
-import Nav from './Nav';
-import Home from './Home';
-import Page2 from './Page2';
+import Nav from './components/Nav';
+import Home from './pages/Home/Home';
+import Profile from './pages/Profile/Profile';
 
 import './App.scss';
 
@@ -13,7 +13,7 @@ function App() {
     <Router>
       <Nav />
       <Route path="/" exact component={Home} />
-      <Route path="/page2/" component={Page2} />
+      <Route path="/profile/" component={Profile} />
     </Router>
   );
 }
