--- src/Page2.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-export default function Page2() {
-  return (
-    <h1>Page 2</h1>
-  );
-}
