+++ src/pages/Profile/Profile.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export default function Profile() {
+  return (
+    <h1>Profile</h1>
+  );
+}
