+++ src/scss/_colors.scss
@@ -1,0 +1,6 @@
+// colors
+$white: #FFF;
+$bg1: #4F5867;
+$bg2: #525B6A;
+$bg3: #656F80;
+$navbar-bg: #F8F8FC;
